feat(alloc): introduce U256 in alloc
